Implement feed silo using Erlangen
…SERVICE implements an Erlangen supervisor that is the new
root of Athens. It spawns a logger, a timer, a feed importer,
and a feed fetcher. The feed fetcher registers events with the
timer to schedule pulling the feeds. It parses the updated
feeds and sends them to the feed importer. The feed importer
is a server that accepts parsed feeds and persists their new
content. The logger is used for... writing logs.

The schema in …STORE is simplified by introducing a timestamp
column to the items table. The log and global-date tables are
obsoleted.

SQL query code is improved. …STORE:RECORD-ITEM is renamed to
INSERT-ITEM.

…STORE:WITH-DATABASE no longer uses pooled connections, and
sets an I/O timeout instead, configured through the :timeout
key. (CCL specific!)